# dept44-example/src/test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: api-service-pet-inventory
  contact: { }
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: 8.0.7-SNAPSHOT
servers:
  - url: http://localhost:58107
    description: Generated server url
tags:
  - name: Pet inventory
    description: Pet inventory operations
paths:
  /pet-inventory-items/{id}/images:
    post:
      tags:
        - Pet inventory
      summary: Add pet inventory item image by id
      operationId: addPetImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: Successful operation
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /pet-inventory-items:
    get:
      tags:
        - Pet inventory
      summary: Get Pet inventory items
      operationId: getPetInventoryList
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PetInventoryItem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /pet-inventory-items/{id}:
    get:
      tags:
        - Pet inventory
      summary: Get Pet inventory item by id
      operationId: getPetInventoryItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetInventoryItem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /pet-inventory-items/{id}/images/{imageId}:
    get:
      tags:
        - Pet inventory
      summary: Get pet inventory item image by id
      operationId: getPetImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "201":
          description: Successful operation
          content:
            '*/*':
              schema:
                type: string
                format: byte
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api-docs:
    get:
      tags:
        - API
      summary: OpenAPI
      operationId: getApiDocs
      responses:
        "200":
          description: OK
          content:
            application/yaml:
              schema:
                type: string
      x-auth-type: None
      x-throttling-tier: Unlimited
      x-wso2-mutual-ssl: Optional
components:
  schemas:
    Problem:
      type: object
      properties:
        detail:
          type: string
        title:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    PetImage:
      type: object
      description: Pet image model
      properties:
        id:
          type: integer
          format: int64
          description: Pet image ID
          examples:
            - 1
        fileName:
          type: string
          description: File name
          examples:
            - image.jpg
        mimeType:
          type: string
          description: Mime type
          examples:
            - image/jpeg
    PetInventoryItem:
      type: object
      description: Pet inventory item model
      properties:
        id:
          type: integer
          format: int64
          description: Pet ID
          examples:
            - 1
        price:
          type: number
          format: float
          description: Pet price
          examples:
            - 1.5
        name:
          type: string
          description: Pet name
          examples:
            - Daisy
        type:
          type: string
          description: Pet type
          examples:
            - DOG
        clientId:
          type: string
          description: Client ID
          examples:
            - joe01doe
        images:
          type: array
          items:
            $ref: "#/components/schemas/PetImage"
  securitySchemes: { }